// src/enum/class.enum.ts
export enum Class {
  Alchemist = "alchemist",
  Artificer = "artificer",
  Barbarian = "barbarian",
  Bard = "bard",
  Cavalier = "cavalier",
  Cleric = "cleric",
  Druid = "druid",
  Fighter = "fighter",
  Gunslinger = "gunslinger",
  Monk = "monk",
  Oracle = "oracle",
  Paladin = "paladin",
  Ranger = "ranger",
  Rogue = "rogue",
  Samurai = "samurai",
  Sorcerer = "sorcerer",
  Swashbuckler = "swashbuckler",
  Summoner = "summoner",
  Warlock = "warlock",
  Witch = "witch",
  Wizard = "wizard"
}